extract windows emulation header of atomic & add attribution
